CoreDisTools: Implement Near-Differ
This change implements the near-diff functionality necessary for
asm-diff analysis. The near-differ compares two code-sections
for equivalency using the LLVM disassembler .

The two code blocks are disassembled in parallel looking for similar
instructions. In the case of constant operands, some heuristics are
applied via the areOffsetsEquivalent() callback to determine if the
constants can be considered equivalent (ex: constant offsets from the
start of the respective code sections).