Fix path setting in bootstrap scripts
The bootstrap scripts are meant to set the PATH to include the path to the
jitutils executables. However, the way the export/set commands are
written, we attempt to add the actual executable to the path, rather than
the path to the executable ($root/jitutils/bin/jit-diff instead of
$root/jitutils/bin). This change updates the command so that when the user
runs the bootstrap scripts, the correct path is added to PATH, and the
user can just run the utilities without using their full path name.