Fix issues in jit-format; download format/tidy
This change updates bootstrap.cmd to download clang-format and clang-tidy
to the bin directory

We introduce the SOURCE_FORMATTING define to the windows compile_commands
database which is used in the pal headers to bypass clang-tidy build
errors.

This change also adds a link to the getstarted.md documentation if
clang-tidy or clang-format fail.

Fix issue with failure mode if clang-format or clang-tidy are not found
and fail gracefully in those cases.