Fix compile_commands.json for non-Windows
On non-Windows platforms, we use CMAKE_EXPORT_COMPILE_COMMANDS, which
creates commands for every file and every project. We are only
interested in jit files, and we need each project to be separated so
that clang-tidy does not attempt to run on each file multiple times. If
it does, it replaces NULL with nullptr several times and we end up with
nullptrptrptr all over the code. This change takes the compile_commands
database generated by CMake, parses it and writes a new compile_commands
database with only the commands we are currently interested in.

Additionally, this change add -style=file to the clang-format invocation
to make sure that clang-format knows to look for the style in the
.clang-format file in the jit directory.