Change how we run clang-format
Command strips out blank lines when capturing standard output from
processes it runs, so we need a workaround for clang-format in order to
use this functionality. Clang-format has an -output-replacement-xml that
we can use which prints the fixes in an xml format. We can parse this
xml, and use it to determine if there are formatting errors. We can also
use it to calculate and print out the changes that would need to be
made, making this workaround a more verbose method.

The new output will look like this:

clang-format: there are formatting errors in
/Users/adiaaida/code/coreclr/src/jit/alloc.cpp
At Line 106 Before:
{
      assert(getDefaultPageSize()  !=  0);
After:
{
    assert(getDefaultPageSize()  !=  0);

At Line 107 Before:
    assert(getDefaultPageSize()  !=  0);
After:
    assert(getDefaultPageSize() !=  0);

At Line 107 Before:
    assert(getDefaultPageSize() !=  0);
After:
    assert(getDefaultPageSize() != 0);

At Line 107 Before:
    assert(getDefaultPageSize() != 0);
  assert(isInitialized());
After:
    assert(getDefaultPageSize() != 0);
    assert(isInitialized());