Fix SetPlatform() and bootstrap for Linux platforms
I assumed that coreclr built into the name of the operating system on
various flavors of Linux, but it turns out they are built into
Linux.<arch>.<build>. This change fixes SetPlatform to set the OS as
Linux if the OS reported by dotnet --info is not Windows or OSX.

Additionally, in bootstrap, we were trying to get the exes which do
not exist for Linux platforms (they do not have extensions). This
change fixes the wget, and makes the files executable. It also only
checks the version if the tools were already on the path.