Rewrite compile commands on Windows
If the user did not specify a compile_commands.json, but one exists, we
will default to using it. This is to allow the use of
compile_commands.json files that were generated by cmake with NMake
Makefiles on Windows. These files require slightly more processing than
cmake files generated by cmake on non-windows platforms, as clang gets
confused by some of the windows specific flags, and we also have to force
the use of various other flags.

This change also modifies the output of DoClangFormat so that the output
doesn't get interleaved during the parallel runs of clang-format. We write
all the output to a list, and then print it out after we have run all
clang-format instances.

This change also modifies the behavior of jit-format if
compile_commands.json does not exist. If none exists, run cmake
usenmakemakefiles (on windows) or cmake configureonly (on non-windows)
to generate the json.

This change also removes the logic to parse build logs