Fix bugs in boostrap
Bootstrap.sh calls validate_url, whose definition had parentheses. This
was not legal in the lab, so this change removes those parentheses.

Additionally, for the lab we need to install the 3.8 clang tools if they
are not yet installed. This change will download 3.8 if another version
of the tools is already installed.