Add superpmi collect automation
This commit includes scripting for collecting superPMI over the coreclr
tests and crossgen of the Framework Assemblies. It also includes the ability
to expand collection over any managed dll passed to the tool. It also includes
a script to upload, download and list the packages that are created by the tool
into Azure Blob Storage.