factor diff logic, add support for --crossgen flag
This change moves all of the diff logic to jit-diff, which will invoke
jit-dasm once for each tag.

jit-dasm now takes an argument that specifies the crossgen executable to
use, and an optional jit argument that will be passed along to crossgen
via /JITPath.

jit-diff has been modified to take an additional argument that specifies
the crossgen executable to use. With this argument, jit-diff will use
clrjit from the the --base and --diff arguments (which are now
directories instead of the full crossgen path) with the specified
crossgen. This should prevent changes to
System.Private.CoreLib.dll (which must match the crossgen version) from
manifesting as jit diffs.