Fix error messages for various scenarios in jit-format
We need better error messages/handling in jit-format. This change
updates the following error scenarios:

1) The specified coreclr path does not exist or is not a wellformed
coreclr repo. We chech to make sure that both the directory exists, and
that build.cmd and build.sh exist within the specified directory.

2) jit-format was cancelled mid-run, specifically before the clang-tidy
step has a chance to delete the newCompileCommands file. We now check to
see if newCompileCommands already exists, and delete it before trying to
move the full compileCommands database to that file. File.Delete will do
nothing if the file does not exist.

Fixes #63.
Fixes #64.