Look for specific dlls instead of matching in jit-diff
With this change, jit-diff will look for crossgen binaries named
"crossgen" or "crossgen.exe", and jit binaries named "clrjit.dll" or
"libclrjit.so". We used to match files with similar names, creating an
issue when the base or diff directories had multiple files with "clrjit"
in the name.

This also adds a minor clarification to the docs (the crossgen
executable has an exe extension on Windows only).