Fix `build -f`
In the migration to csproj, the frameworks-populating of `build -f` was
lost, because packages.csproj can't specify standalone like project.json
could.  With csproj, an explicit --runtime argument must be passed to
dotnet restore in order to publish as a standalone.  The dotnet publish
--runtime in turn requires that the project has been restored for the
specified runtime, so add an explicit `dotnet restore --runtime` in the
build scripts when processing -f.