Finish all diff tasks; don't exit early on failure (#95)
If an asm run fails, e.g., in the baseline run, we were
immediately exiting. Instead, run all the baseline, diff,
and analyze tasks, so you can get as much information as
possible from the code you've got.

Finally, display count of errors generating base/diff at
the end of the output, so the failure messages don't
get lost in the progress output.