Improve jitutils
Short summary:

1. Improved usage model for jit-diff, especially, add support for much simpler
   command lines using better defaults. Read the rewritten diffs.md
   for details.
2. Improved jit-diff parallelism when diffing the test tree. (Previously,
   dasm generation within a recursive directory walk was not parallel.)
3. Changed jit-diff argument parsing to output the usage message on (most)
   errors (namely, the errors we validate, not the errors the argument
   library validates. It would be nice to do both.).
4. Fixed many crashing bugs

Details:

Using jit-diff is cumbersome due to the need to specify so many
long paths. These can be specified in the config.json file in
the directory pointed to by the `JIT_UTILS_ROOT` environment
variable, but that is inflexible.

This change introduces an alternate way to specify all the required
data for jit-diff, by detecting or specifying the dotnet/coreclr
repo root, and automatically finding all the paths that are required.

The defaults can be altered by using the new `--base_root`, `--diff_root`,
`--arch`, and `--build` arguments, or their respective config.json settings.

See the rewritten diffs.md document for much more detail and various
examples of using this.

Also:
1. To generate asm diffs for the test tree, specify the new "--tests"
argument, as well as the "--test_root" argument (if "--test_root" can't
be determined using the defaults).
2. jit-diff now automatically runs jit-analyze. It now has a "--noanalyze"
option to prevent this.
3. jit-diff now takes an optional "--gcinfo" argument (which it passes
through to jit-dasm) to specify emitting GC info into the generated dasm
files.
4. jit-format automatically detects the repo root, if run with the
current directory within the dotnet/coreclr repo. So, the "--core_root"
argument is no longer required, if the repo root can be found.
5. Rewrote "getting started" document to account for new defaults,
and other changes.
6. Move all cijobs documentation to cijobs.md.
7. Remove `--tag` option from `jit-dasm`: simply pass the `--output` path you desire.
8. Support diff'ing System.Private.CoreLib.dll or the frameworks assemblies, in addition to
   the benchmarks or the full test tree. I regularized this, and added a lot of missing
   argument checking. I also renamed the arguments since "only" is incorrect. So, now
   there is `--corelib`, `--frameworks`, `--tests`, and `--benchmarks`.
9. Removed `jit-diff install --last_successful` option because it didn't work.
10. Removed the two old sample json files (that didn't work), and replace them with
    a sample config.json file that does.