Support netcoreapp2.0
Adds a <TargetFrameworks> element to each project that includes
netcoreapp2.0 as a target. This is useful because it makes it possible
to build and run jitutils with a 2.0 install of the cli, which doesn't
come with a 1.0 runtime.

For now, the default target is still netcoreapp1.0, to prevent
breaking ci jobs that run jitutils with a 1.0 runtime.