Show when number of methods differs in base/diff (#122)
When multiple same-named methods exist in different numbers in base and
diff, make it clearer in the output. Also call out methods that are exclusive
to base or diff.

Allow analysis to skip reconciliation; this option is useful when
the larger diffs are all in unique methods and you want to focus in on the
methods that are common.