PMI: fix DRIVEALL for use with dotnet cli
Details of creating a cloned process are a bit different for the CLI.
Handle this via condtionally compiled code for now.

Still other changes would be required to support mixing the "host" and
target runtimes. It might be worth explicitly specifying these somehow.

Also did some folding of long lines.