PMI: rework as visitor pattern to support generic type exploration (#128)
This change adds the ability for PMI to create instantations of some
generic types (limited to types with one or two parameters only for now).

Because of constraints it is not possible to easily tell in advance how
many instantations will result. Also in the future we will want to do
something similar for generic methods, and may want to produce different
kinds of output markup.

So this change also refactors the PMI code to use a visitor pattern for
traversing types and methods. Three concrete visitors are implemented:
one for counting types and methods, one for jitting a range of methods, and
one for jitting just a single method.

The actual strategy for instantiating generics is held in the `Worker` class.
We attempt up to 4 instantations of single parameter types and 8 of two
parameter types, with a mixture of types. There is room for enhancement here,
both in trying out suitable struct types and in looking for types that will
satisfy various constraints. A limited amount of constraint validation is
done up front.

Aside from extra methods and types seen via instantations this should be
functionally equivalent to the base version.