PMI: start keeping track of missing and failing things (#132)
Start keeping tabs on assemblies that won't load, types that can't
be instantiated, and methods that can't be instantiated.

Note this does not track methods that fail to jit. Need to think
about that some more.

Add provisional support for running PMI on all the assemblies in a
directory tree (core only for now). Report aggregate stats if we end
up PMIing multiple assemblies.

Convert remaining formatting in pmi.cs to interpolated style.