Fix some concurrency issues.
Name marker files after the corresponding assembly to avoid name
collisions when multiple PMI processes are running concurrently.

Add hidden option to tell jit-dasm-pmi it doesn't need to copy in
the jit to test because some higher level command has taken care of it.

Orchestrate jit copies for PMI diffs in jit-diff and tell jit-dasm-pmi
not to do jit copies.