Initial support for pmi based asm diffs (#134)
Add jit-pmi-dasm as a new tool. Add plumbing for
this tool to jit-diffs as `pmidiff`, so one can now say:

  `jit-diff diff --pmi  --base --base_root d:\repos\coreclr2 --diff`

Output is compatible with existing `jit-analyze`.

Under the covers this invokes `jit-dasm-pmi` which in turn
invokes `pmi`.

Details:

Because we're not (yet?) using altjits for same-arch pmi, we have
to copy the jit under test into the test overlay dir. This means
carefully saving off the jit that was there and putting it back when
done, as the primary jit name can't be altered. Also we can't run
base and diff jit tasks at the same time. So we run one group, then
the other.

Add hidden option to tell `jit-dasm-pmi` it doesn't need to copy in
the jit to test because some higher level command has taken care of it.
Orchestrate jit copies for PMI diffs in `jit-diff` and tell `jit-dasm-pmi`
not to do jit copies.

Don't bother with marker files when `pmi` is driven by `jit-pmi-dasm`.
This is a bit of a hack but I haven't found the right way to generate
marker files that is robust in the presence of concurrent PMIs.

Update `pmi` to newline after new method banner so the jit dasm output
starts on its own line.

Quiet down the default output for `jit-diff diff`. Can restore it to its full
glory with `--verbose`. Also allow specifying a particular assembly to diff.