Some minor usability tweaks in jitutils
Have `jit-diff` pass a description string down to `jit-analyze` so the
summary can describe what was being diffed. Capture the arch that was
actually used and add it to the description.

Have `jit-analyze` accept that description string and display it in the
summary.

Remove a diagnostic "adding" message inadvertenly left in `jit-analyze`'s
processing of files with text diffs.

Don't show final file diff (but no size diff) message in the analysis
summary if there were no such files (which is commonly the case).

Update Usage for `pmi` to indicate you must pass `pmi.dll` as the first arg
to your launcher if you launched via `dotnet` or `corerun`. This should
partially address #140.

Add a bit more EH to `jit-dasm-pmi` in a attempt to reduce the volume of
dotnet failure popup dialogs (seen when altjit asserts cause crashes). More
is needed here to make this fully robust.