PMI: Improve constraint checking logic
The logic to check for generic parameter type constraints was wrong.
If the constraint it itself generic then just skip checking and let
the runtime figure it out.

Handle the other classes of special constraint. Fix the check for the
class/interface constraint.

Add `long` to the types we try and use for instantiations since it
stresses x86.