Allow --test-root to point to an arbitrary tree
If you point at an arbitrary tree, and pass --tests, then generate
asm for all assemblies in the tree. This allows you to generate
asm for just a subset of the test tree by passing that as the root.

Whether we are looking at the test root or an arbitrary subtree is
determined by looking for the test tree root subdirectories we
expect to find (JIT and Interop).