Force the new project system for cs projects.
Otherwise it does not see `TargetFramework` from the props file and uses the old one that doesn't show *.cs files in VS (because that are not included by default in the old version), but it also doesn't allow to include them because then you will hit build errors (one file is included twice).
See the issue description https://github.com/dotnet/project-system/issues/1358 and workaround is in the end.