Have jit-format explicitly specify that it needs a compile_commands.json from the CMake configuration
Right now the fact that jit-format needs a compile_commands.json file isn't explicit anywhere in CoreCLR and nowhere in CoreCLR uses the compilation database. This change makes jit-format explicitly request the compile_commands.json on Unix systems where there is a `-cmakeargs` parameter to our build scripts.

There's no comparable flag on our Windows build scripts, so there's not a good way to move that knowledge into jit-format. As a result, I'm planning a PR into CoreCLR that will only tell CMake to generate a compile_commands.json when the `usenmakemakefiles` parameter is passed (which makes sense because the only reason to use the NMake generator is to generate a compile_commands.json file).