MutateTest: experimental tool to mutate existing test cases
The MutateTest tool takes existing test cases and adds various syntactic
mutations that alter the program expression and execution without altering
the computed results. The aim is to extend the usefulness of existing tests
and to stress the jit in novel ways, for instance, introducing EH constructs
into methods.

Mutation is done at the syntax tree level. There is a suite of basic mutations
and a number of compositional operators to combine or alternate mutations. For
example a block can be wrapped in the try of a try/catch, the try of a
try/finally, the catch of a try/catch where the try throws an exception that
gets caught, and so on.

Some care must be taken when blocks are moved into finallys or catches because
execution of blocks in these contexts can require considerably more stack. The
mutator tries to avoid causing inadvertent stack overflow both but the current
strategies are only partially successful. So there's also an exclusion list to
bypass mutation of some problematic test cases.

Current help text:
```
Mutate:
  Take an existing test case and produce new test cases via mutation

Usage:
  Mutate [options] <InputFile>

Arguments:
  <InputFile>    Input test case file or directory (for --recursive)

Options:
  --ehStress                 add EH to methods
  --structStress             replace locals with structs
  --showResults              print modified programs to stdout
  --verbose                  describe each transformation
  --quiet                    produce minimal output
  --recursive                process each file recursively
  --seed <seed>              random seed
  --stopAtFirstFailure       stop each test at first failure
  --emptyBlocks              transform empty blocks
  --sizeLimit <sizelimit>    don't process programs larger than this size
  --timeLimit <timelimit>    don't stress programs where compile + run takes more than this many milliseconds
  --projects                 look for .csproj files instead of .cs files when doing recursive exploration
  --onlyFailures             only emit output for cases that fail at runtime
  --version                  Display version information
```