Update jit-analyze to handle multiple metrics (#223)
Alternative take to #215 where jit-analyze is updated to look for
and process multiple bits of metric data. Default is still code size,
alternatives can be specified via `--metric <metricname>`

With the pattern shown here it should be simple to add additional
metrics like the ones we already have.

Updated `jit-diff` to also support the `--metric` option.