One more PMI determinism fix. (#253)
ArrayPool<> is used in several places during the pmi execution.
In particular, ArrayPool<ReadOnlyMemory<>>.Shared.Return is called.
That method registers Gen2GcCallbackFunc on Gen2GcCallback object.
Gen2GcCallbackFunc is then called from Gen2GcCallback finalizer after
a gc. At that point Gen2GcCallbackFunc is jitted. Since gc time is not
deterministic, the disasm output for Gen2GcCallbackFunc and methods
it calls was not deterministic.

The fix is to force the registration of Gen2GcCallbackFunc and then
force a gc so that Gen2GcCallbackFunc is jitted.