Undo a non-determinism workaround.
I added a workaround for disassembly non-determinism in #255.
See #255 for details of what causes the non-determinism.
It turns it relied on a reflection hole that allowed
init-only static fields to be modified after static constructor has
been called. That hole was fixed in https://github.com/dotnet/runtime/pull/37849
so the workaround is no longer valid and causes an exception from pmi.

I don't see a way to work around the non-determinism without changing
framework code so for now I'm just reverting the workaround. Unfortunately, that
means that we can get non-deterministic disassembly for any method that inlines
System.Threading.Thread.GetCurrentProcessorId.

Fixes #271.