Remove explicit "src" level from source path construction
I believe this to be the minimum change to facilitate removal of
the duplicated "src" folder level under "src/coreclr" in the
runtime repo. Before the removal we'll need to make a coordinated
runtime change to pass the source directory as "src/jit" instead
of just "jit"; the removal PR should subsequently change the
"src/jit" folder as the input argument to jit-format to just "jit".

Thanks

Tomas