PMI fixes
1. Handle a special case in DRIVEALL when PMI is invoked using the `pmi.exe` tool,
instead of `dotnet.exe` or `corerun.exe`, where we need to remove the
implicit `pmi.dll` argument before spawning the child process.
2. In DRIVEALL, when preparing to spawn the sub-process, we replace the DRIVEALL
command with PREPALL. This replacement needs to be case-insensitive, or else a
command-line like "pmi.exe DriveAll" will fail.
3. Change PrepOne to behave the same as PrepAll w.r.t. the `-QUIET` option.
4. Update the help message with `-Tailcalls` and `-Inlines` options, and note that
for PrepOne, the method index is optional (defaults to zero).
5. Fixed (the weird) arg parsing to search for the leading dashes to the options
(before, `PrepOne-Tailcalls` was treated as `PrepAll-Tailcalls` since it just searched
for `all`).