Add option to enable Linux cross builds for jit-format.
Currently, when jit-format attempts to build a fresh version of
compile-commands.json on Linux, it will always attempt a cross
build, regardless of whether this is accurate for the host system.

This patch adds an option to enable cross builds when on Linux,
rather than defaulting to always building with -cross.