Add instructions retired explorer to jitutils
Seems like it makes sense to include this here rather than keeping
it in my own private repo.

We might consider renaming it, since it only optionally looks for
instructions retired; mainly its job is to parse ETW and summarize
the profile data and the jit time information.

It also knows about BDN intervals and so is especially useful for
summarizing profiles gathered with BDN's `-p ETW` option.