Update coredistools (#370)
1. Update from LLVM 13.0.1 to 17.0.6
2. Change Linux/Mac build to build llvm-tblgen from source instead of
downloading a pre-built version. LLVM doesn't always publish all
architecture versions of this tool.
3. Change Linux to build with CBL-Mariner container.

TO DO:
- [x] Mac: properly download built llvm-tblgen artifact and put it on
the path so the build-coredistools step can find it.
- [x] Linux: fix the CBL-Mariner build. Mariner currently has clang12
and is expected to update to clang16 soon. Will that fix problems? Or,
back off to ubuntu again, and stop doing cross-compiler builds for
linux-x64 (like we do for Mariner).

Fixes #372