MLCSE: lay groundwork for optimizing for size (#399)
Add a new `--optimizeSize` option where PolicyGradient tries to reduce
code size instead of perf score.

Since there are now two possible optimization objectives, we have the
potential for trading off one for the other. Update MCMC to track the
"pareto frontier" for the methods it explores and save the data for
visualization.

Contributes to dotnet/runtime#92915.