Add updated R2R analysis
We no longer get method load events for R2R methods, but there are events
that are fired when the runtime goes to find an R2R entry point that are
almost as good. Enable these via `-RuntimeLoading` command line arg to perfview.

Add support for parsing these events and inferring the length of R2R methods
from them.